+++ CHANGELOG.md
@@ -19,6 +19,8 @@
   ### :bug: Bug Fix
   - `divider`
     - [#511](https://github.com/wix-incubator/rich-content/pull/511) divider's width set to 100% in editor&viewer when no initial state of width is given
+  - `divider`
+    - [#525](https://github.com/wix-incubator/rich-content/pull/525) fix emoji delete bug
   ### :house: Internal
   - `editor`
     - [#510](https://github.com/wix-incubator/rich-content/pull/510) custom icons code refactoring